requirements.txt
yt-dlp
google-api-python-client
rich
isodate
pyinstaller
imageio-ffmpeg
pyfiglet
mutagen
numpy
pyaudio
speedtest-cli
